Add GetMoviesWithPostersUseCase
