+++ feature/main/src/main/java/com/chaeny/moviechart/MainViewModel.kt
@@ -3,14 +3,8 @@ package com.chaeny.moviechart
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.chaeny.moviechart.mapper.MovieIdMapper
 import com.chaeny.moviechart.repository.GetMoviesResult
-import com.chaeny.moviechart.repository.KobisRepository
-import com.chaeny.moviechart.repository.TmdbRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.async
-import kotlinx.coroutines.awaitAll
-import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
@@ -18,9 +12,7 @@ import javax.inject.Inject
 
 @HiltViewModel
 internal class MainViewModel @Inject constructor(
-    private val kobisRepository: KobisRepository,
-    private val tmdbRepository: TmdbRepository,
-    private val movieIdMapper: MovieIdMapper
+    private val getMoviesWithPostersUseCase: GetMoviesWithPostersUseCase
 ) : ViewModel() {
 
     private val _movies = MutableStateFlow<List<Movie>>(emptyList())
@@ -45,16 +37,16 @@ internal class MainViewModel @Inject constructor(
         _movies.value = emptyList()
         _isLoading.value = true
         viewModelScope.launch {
-            val result = kobisRepository.getMovies(_selectedTab.value)
+            val result = getMoviesWithPostersUseCase(_selectedTab.value)
             handleMoviesResult(result)
             _isLoading.value = false
         }
     }
 
-    private suspend fun handleMoviesResult(result: GetMoviesResult) {
+    private fun handleMoviesResult(result: GetMoviesResult) {
         when (result) {
             is GetMoviesResult.Success -> {
-                val moviesWithPosters = loadMoviePosters(result.movies)
+                val moviesWithPosters = result.movies
                 _movies.value = moviesWithPosters
             }
             is GetMoviesResult.NoResult -> {
@@ -69,18 +61,6 @@ internal class MainViewModel @Inject constructor(
                 Log.e("MainViewModel", "NetworkError")
                 _movies.value = emptyList()
             }
-        }
-    }
-
-    private suspend fun loadMoviePosters(movies: List<Movie>): List<Movie> {
-        return coroutineScope {
-            movies.map { movie ->
-                async {
-                    val tmdbId = movieIdMapper.getTmdbId(movie.id)
-                    val posterUrl = tmdbRepository.getPosterUrl(tmdbId)
-                    movie.copy(posterUrl = posterUrl)
-                }
-            }.awaitAll()
         }
     }
 }
